refactored fetch requests into separate method
